+++ modules/hanna-react/CHANGELOG.md
@@ -5,13 +5,14 @@
 - ... <!-- Add new lines here. -->
 - feat: Add component `ReykjavikWaves`
 - `DropdownButton`:
-  - feat: Add prop `Toggler` to for custom toggler content
+  - feat: Add prop `Toggler` for custom toggler content
   - feat: Add "customitem" object with `Content`, `modifier` and `current`
 - `MainMenu2`:
   - feat: Add prop `redhot` to `MainMenu2Props.items.hot` items to always show
     them on mobile
   - feat: Add "customitem" object with `Content`, `modifier` and `current`
   - docs: Minor JSDoc corrections for on-click handlers
+- fix: Clicking `.SearchInput__button` always passed empty string as value
 
 ## 0.10.160